Added some new instructions for peeking into the stack and transferring its contents to a data register, added PANIC instruction for panicking threads easily and getting some useful logs for debugging.
NTHRD is now 0x1A instead of 0x20, because it was jumping six entire digits.

SWAP is for swapping the last two elements of the stack.

Also removed 0x13 and 0x14 AKA Print Anything from Data and Print Anything from Stack. Now the Immediates have the Display trait, so it's to be considered deprecated.